+++ pkg/utils/sort-files-by-date.go
@@ -1,4 +1,4 @@
-package usbsync
+package utils
 
 import "os"
